refactor(rlm): extract query loop compiler
Amp-Thread-ID: https://ampcode.com/threads/T-01a039e2-f157-7753-9131-d75c8b31f3c9
